requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Pillow==10.0.1
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.0.3
paddlepaddle==2.5.1
numpy==1.24.3
opencv-python==4.8.1.78
scikit-image==0.21.0
python-magic==0.4.27